feat: updated auth layout handling
