+++ apps/firebase-chat/src/App.vue
@@ -1,20 +1,24 @@
 <script setup lang="ts">
+import { Suspense } from 'vue';
 import useAuthStore from '@store/auth';
 
-const store = useAuthStore();
+useAuthStore();
 </script>
 
 <template lang="pug">
 q-layout(view="lHh Lpr lff")
   q-page-container.main-container
-    router-view(v-slot="{ Component }")
-      transition(
-        mode="out-in"
-        enter-active-class="animated fadeIn"
-        leave-active-class="animated fadeOut"
-      )
-        Component(:is="Component")
-    q-btn(v-if="store?.user" @click.prevent="store.logout") Logout
+    Suspense
+      template(#default)
+        router-view(v-slot="{ Component }")
+          transition(
+            mode="out-in"
+            enter-active-class="animated fadeIn"
+            leave-active-class="animated fadeOut"
+          )
+            Component(:is="Component")
+      template(#fallback)
+        p Loading...
 </template>
 
 <style>
