+++ apps/firebase-chat/src/layouts/default.vue
@@ -4,8 +4,13 @@ q-layout
 </template>
 
 <script setup lang="ts">
+// import useAuthStore from '@store/auth';
 import { useRoute } from 'vue-router';
+import useAuthStore from '@store/auth';
 
 const route = useRoute();
 route.meta.requiresAuth = true;
+
+const authStore = useAuthStore();
+await authStore.userLoad();
 </script>
