+++ apps/firebase-chat/src/layouts/unauthenticated.vue
@@ -2,3 +2,11 @@
 q-layout
   router-view
 </template>
+
+<script lang="ts">
+import useAuthStore from '@store/auth';
+
+const store = useAuthStore();
+
+await store.userLoad();
+</script>
